+++ src/main/java/Backend/FinalProject/Tool/Validation.java
@@ -1,0 +1,47 @@
+package Backend.FinalProject.Tool;
+
+import Backend.FinalProject.domain.Member;
+import Backend.FinalProject.dto.ResponseDto;
+import Backend.FinalProject.repository.MemberRepository;
+import Backend.FinalProject.sercurity.TokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import javax.servlet.http.HttpServletRequest;
+
+@RequiredArgsConstructor
+@Service
+public class Validation {
+
+    private final TokenProvider tokenProvider;
+    private final MemberRepository memberRepository;
+
+    // RefreshToken 유효성 검사
+    @Transactional
+    public Member validateMember(HttpServletRequest request) {
+        if (!tokenProvider.validateToken(request.getHeader("RefreshToken"))) {
+            return null;
+        }
+        return tokenProvider.getMemberFromAuthentication();
+    }
+
+    // Token 유효성 검사
+    public ResponseDto<?> validateCheck(HttpServletRequest request) {
+
+        // RefreshToken 및 Authorization 유효성 검사
+        if (request.getHeader("Authorization") == null || request.getHeader("RefreshToken") == null) {
+            return ResponseDto.fail("NEED_LOGIN", "로그인이 필요합니다.");
+        }
+        Member member = validateMember(request);
+
+        // 토큰 유효성 검사
+        if (member == null) {
+            return ResponseDto.fail("INVALID TOKEN", "Token이 유효하지 않습니다.");
+        }
+        return ResponseDto.success(member);
+    }
+
+
+
+}
