+++ src/main/java/Backend/FinalProject/service/MyPageService.java
@@ -1,6 +1,7 @@
 package Backend.FinalProject.service;
 
 import Backend.FinalProject.Tool.Time;
+import Backend.FinalProject.Tool.Validation;
 import Backend.FinalProject.domain.Application;
 import Backend.FinalProject.domain.Member;
 import Backend.FinalProject.domain.Post;
@@ -30,13 +31,14 @@ public class MyPageService {
     private final PostRepository postRepository;
 
     private final WishListRepository wishListRepository;
+    private final Validation validation;
 
     Time time = new Time();
 
     // 신청한 모임 조회
     public ResponseDto<?> application(HttpServletRequest request) {
 
-        ResponseDto<?> responseDto = validateCheck(request);
+        ResponseDto<?> responseDto = validation.validateCheck(request);
 
         if (!responseDto.isSuccess()) {
             return responseDto;
@@ -56,7 +58,7 @@ public class MyPageService {
     // 참여 중인 모임 조회
     public ResponseDto<?> participation(HttpServletRequest request) {
 
-        ResponseDto<?> responseDto = validateCheck(request);
+        ResponseDto<?> responseDto = validation.validateCheck(request);
 
         if (!responseDto.isSuccess()) {
             return responseDto;
@@ -113,7 +115,7 @@ public class MyPageService {
     // 내가 주최한 모임 조회
     public ResponseDto<?> leader(HttpServletRequest request) {
 
-        ResponseDto<?> responseDto = validateCheck(request);
+        ResponseDto<?> responseDto = validation.validateCheck(request);
 
         if (!responseDto.isSuccess()) {
             return responseDto;
@@ -149,7 +151,7 @@ public class MyPageService {
 
     // 찜한 모임 조회
     public ResponseDto<?> addWish(HttpServletRequest request) {
-        ResponseDto<?> responseDto = validateCheck(request);
+        ResponseDto<?> responseDto = validation.validateCheck(request);
 
         if (!responseDto.isSuccess()) {
             return responseDto;
@@ -175,31 +177,4 @@ public class MyPageService {
         return ResponseDto.success(wishListDto);
 
     }
-
-
-        // 토큰
-        public Member validateMember (HttpServletRequest request){
-            if (!tokenProvider.validateToken(request.getHeader("RefreshToken"))) {
-                return null;
-            }
-            return tokenProvider.getMemberFromAuthentication();
-        }
-        // 토큰이 있는지 보고 로그인 여부를 판단하는 메소드
-        private ResponseDto<?> validateCheck (HttpServletRequest request){
-
-            // RefreshToken 및 Authorization 유효성 검사
-            if (request.getHeader("Authorization") == null || request.getHeader("RefreshToken") == null) {
-                return ResponseDto.fail("NEED_LOGIN", "로그인이 필요합니다.");
-            }
-            Member member = validateMember(request);
-
-            // 토큰 유효성 검사
-            if (member == null) {
-                return ResponseDto.fail("INVALID TOKEN", "Token이 유효하지 않습니다.");
-            }
-            return ResponseDto.success(member);
-        }
-
-
-
 }
